Title: 길이가 무한대인 수평선 트랙, Time: 67ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250609/길이가 무한대인 수평선 트랙/horizontal-track-with-infinite-length.cpp
@@ -11,14 +11,14 @@ int N, T;
 int main() {
     
     cin >> N >> T;
-    set<int> s;
+    set<long long> s;
 
     while(N--){
 
         int a,b;
         cin >> a >> b;
         // T초후 최종 위치
-        int target = a + b * T;
+        long long target = a + b * T;
         while(1){
             // target보다 크거나 같은 위치 모두 제거
             // 앞지른 사람은 제거하기 위해
